Bump require for new AmplNLWriter release (#19)
* Bump require for new AmplNLWriter release

* Int64 -> Int

* [ci skip] remove coveralls

+++ src/solvers/SYMPHONY_MPS.jl
@@ -31,6 +31,6 @@ end
 function parse_solution!(::NEOSSolver{:SYMPHONY, :MPS}, m::MPSModel)
 	for v in matchall(r"V(\d+)\W+(-?[\d.]+)", m.last_results)
 		regmatch = match(r"V(\d+)\W+(-?[\d.]+)", v)
-		m.solution[parse(Int64, regmatch.captures[1])] = parse(Float64, regmatch.captures[2])
+		m.solution[parse(Int, regmatch.captures[1])] = parse(Float64, regmatch.captures[2])
 	end
 end
